+++ rust/info-fetcher-commons/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod cache;
 pub mod config;
 pub mod http_error;
 pub mod telemetry;